requirements.txt
flask
google-genai
twilio
gunicorn